Revert "demux/packet: fix demux_packet_shorten"
This reverts commit e81521d81bf9f8895a69c8c23c7a22b6a3458e03.

This change shouldn't be needed, and in fact it's wrong. The FFmpeg API
function could do anything it wants with the packet, including changing
the packet data pointer. Likewise, it's not guaranteed that the
referenced packet's fields mirror the current state of the mpv packet
struct (the AVPacket is only kept for the AVBuffer and the side data
stuff).